refactor: move module to it's own separate crate
