+++ crates/module/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "module"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0.195", features = ["derive", "alloc", "rc"] }
+uuid = { version = "1.6.1", features = ["v4", "serde"] }
