+++ crates/project/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 uuid = { version = "1.6.1", features = ["v4", "serde"] }
 utils = { path = "../utils" }
+module = { path = "../module" }
 serde = { version = "1.0.195", features = ["derive", "alloc", "rc"] }
 serde_json = "1.0.111"
 erased-serde = "0.4.2"
